+++ src/gemseo_benchmark/problems/base_problem_configuration.py
@@ -356,8 +356,8 @@ class BaseProblemConfiguration(metaclass=ABCGoogleDocstringInheritanceMeta):
         file_path: str | Path = "",
         infeasibility_tolerance: float = 0.0,
         max_iteration_number: int = 0,
-        plot_kwargs: Mapping[str, ConfigurationPlotOptions] = READ_ONLY_EMPTY_DICT,
-        grid_kwargs: Mapping[str, str] = READ_ONLY_EMPTY_DICT,
+        plot_settings: Mapping[str, ConfigurationPlotOptions] = READ_ONLY_EMPTY_DICT,
+        grid_settings: Mapping[str, str] = READ_ONLY_EMPTY_DICT,
         use_iteration_log_scale: bool = False,
     ) -> None:
         """Compute the data profiles of given algorithms.
@@ -371,9 +371,9 @@ class BaseProblemConfiguration(metaclass=ABCGoogleDocstringInheritanceMeta):
             infeasibility_tolerance: The tolerance on the infeasibility measure.
             max_iteration_number: The maximum number of iterations to plot.
                 If ``0``, this value is inferred from the longest history.
-            plot_kwargs: The keyword arguments of `matplotlib.axes.Axes.plot`
+            plot_settings: The keyword arguments of `matplotlib.axes.Axes.plot`
                 for each algorithm configuration.
-            grid_kwargs: The keyword arguments of `matplotlib.pyplot.grid`.
+            grid_settings: The keyword arguments of `matplotlib.pyplot.grid`.
             use_iteration_log_scale: Whether to use a logarithmic scale
                 for the number of iterations axis.
         """
@@ -395,8 +395,8 @@ class BaseProblemConfiguration(metaclass=ABCGoogleDocstringInheritanceMeta):
         data_profile.plot(
             show=show,
             file_path=file_path,
-            plot_kwargs=plot_kwargs,
-            grid_kwargs=grid_kwargs,
+            plot_settings=plot_settings,
+            grid_settings=grid_settings,
             use_evaluation_log_scale=use_iteration_log_scale,
         )